refactor(studio): extract speaker/room, audio and input command modules
Phase 2, continuing the studio/main.rs command extraction. Moves three more
themed groups out of main.rs, following the established pattern (no behaviour
change: command names and signatures unchanged, so OSC addresses and the
frontend invoke contract are identical):

- commands/speakers.rs — 24 commands: room ratios, per-speaker geometry
  (position, delay, spatialization, crossover, name) and layout add/remove/move/
  apply.
- commands/audio.rs — 5 commands: output sample rate, output device, refresh,
  audio config apply.
- commands/input.rs — 15 commands: bridged-input config and the live
  device/backend path (node, layout, channels, format, mapping, clock, LFE),
  including the layout-file import (it depends only on public helpers).

Also drops the now-unused `fs` import from main.rs (its only user moved into
commands/input.rs).

main.rs drops from 2166 to 1622 lines (2690 -> 1622 across the Phase 2 PRs so
far, ~40%). Tauri app builds; fmt clean.
