+++ omniphony-studio/src-tauri/src/main.rs
@@ -11,7 +11,7 @@ mod osc_parser;
 use std::path::PathBuf;
 use std::sync::atomic::{AtomicI32, Ordering};
 use std::sync::{Arc, Mutex};
-use std::{fs, fs::File, process::Command as ProcessCommand, process::Stdio};
+use std::{fs::File, process::Command as ProcessCommand, process::Stdio};
 
 use app_state::AppState;
 use config::{load_config, save_config, OscConfig};
@@ -23,8 +23,11 @@ use tokio::sync::mpsc::UnboundedSender;
 
 // Tauri command handlers grouped into themed modules under `commands/`. Glob-
 // imported so `generate_handler!` can keep referring to them by bare name.
+use commands::audio::*;
+use commands::input::*;
 use commands::render::*;
 use commands::resampling::*;
+use commands::speakers::*;
 
 // ── shared state wrapper ──────────────────────────────────────────────────
 
@@ -512,278 +515,6 @@ fn control_distance_diffuse_curve(state: State<SharedState>, value: f32) {
 }
 
 #[tauri::command]
-fn control_room_ratio(state: State<SharedState>, width: f32, length: f32, height: f32) {
-    let w = width.max(0.01);
-    let l = length.max(0.01);
-    let h = height.max(0.01);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloats3 {
-            address: "/omniphony/control/room_ratio".to_string(),
-            a: w,
-            b: l,
-            c: h,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_room_ratio_rear(state: State<SharedState>, value: f32) {
-    let v = value.max(0.01);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/room_ratio_rear".to_string(),
-            value: v,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_room_ratio_lower(state: State<SharedState>, value: f32) {
-    let v = value.max(0.01);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/room_ratio_lower".to_string(),
-            value: v,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_room_ratio_center_blend(state: State<SharedState>, value: f32) {
-    let v = value.clamp(0.0, 1.0);
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendFloat {
-            address: "/omniphony/control/room_ratio_center_blend".to_string(),
-            value: v,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_layout_radius_m(state: State<SharedState>, value: f32) {
-    let v = value.max(0.01);
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "radiusM": v }),
-    );
-}
-
-#[tauri::command]
-fn control_layout_config(state: State<SharedState>, payload: serde_json::Value) {
-    send_json_control(&state.osc_tx, "/omniphony/control/config/layout", payload);
-}
-
-#[tauri::command]
-fn control_layout_config_apply(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/config/layout/apply".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_speakers_config(state: State<SharedState>, payload: serde_json::Value) {
-    send_json_control(&state.osc_tx, "/omniphony/control/config/speakers", payload);
-}
-
-#[tauri::command]
-fn control_speaker_az(state: State<SharedState>, id: i32, value: f32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "azimuth": value }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_el(state: State<SharedState>, id: i32, value: f32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "elevation": value }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_distance(state: State<SharedState>, id: i32, value: f32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "distance": value.max(0.01) }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_x(state: State<SharedState>, id: i32, value: f32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "x": value.clamp(-1.0, 1.0) }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_y(state: State<SharedState>, id: i32, value: f32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "y": value.clamp(-1.0, 1.0) }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_z(state: State<SharedState>, id: i32, value: f32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "z": value.clamp(-1.0, 1.0) }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_coord_mode(state: State<SharedState>, id: i32, value: String) {
-    let normalized = if value.trim().eq_ignore_ascii_case("cartesian") {
-        "cartesian"
-    } else {
-        "polar"
-    };
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "coordMode": normalized }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_delay(state: State<SharedState>, id: i32, delay_ms: f32) {
-    let v = delay_ms.max(0.0);
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/speakers",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "delayMs": v }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_spatialize(state: State<SharedState>, id: i32, spatialize: i32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "spatialize": spatialize != 0 }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_name(state: State<SharedState>, id: i32, name: String) {
-    let trimmed = name.trim();
-    if trimmed.is_empty() {
-        return;
-    }
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "name": trimmed }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_freq_low(state: State<SharedState>, id: i32, freq_low: f32) {
-    let value = if freq_low > 0.0 {
-        serde_json::Value::from(freq_low)
-    } else {
-        serde_json::Value::Null
-    };
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "freqLow": value }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speaker_freq_high(state: State<SharedState>, id: i32, freq_high: f32) {
-    let value = if freq_high > 0.0 {
-        serde_json::Value::from(freq_high)
-    } else {
-        serde_json::Value::Null
-    };
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "speakerEdits": [{ "id": id.max(0), "freqHigh": value }] }),
-    );
-}
-
-#[tauri::command]
-fn control_speakers_apply(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/config/layout/apply".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_speakers_add(
-    state: State<SharedState>,
-    name: String,
-    azimuth: f32,
-    elevation: f32,
-    distance: f32,
-    spatialize: i32,
-    delay_ms: f32,
-) {
-    let n = if name.trim().is_empty() {
-        "speaker"
-    } else {
-        name.trim()
-    };
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({
-            "addSpeaker": {
-                "name": n,
-                "azimuth": azimuth,
-                "elevation": elevation,
-                "distance": distance.max(0.01),
-                "spatialize": spatialize != 0,
-                "delayMs": delay_ms.max(0.0)
-            }
-        }),
-    );
-    control_speakers_apply(state);
-}
-
-#[tauri::command]
-fn control_speakers_remove(state: State<SharedState>, index: i32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "removeSpeaker": index.max(0) }),
-    );
-    control_speakers_apply(state);
-}
-
-#[tauri::command]
-fn control_speakers_move(state: State<SharedState>, from: i32, to: i32) {
-    send_json_control(
-        &state.osc_tx,
-        "/omniphony/control/config/layout",
-        serde_json::json!({ "moveSpeaker": { "from": from.max(0), "to": to.max(0) } }),
-    );
-    control_speakers_apply(state);
-}
-
-#[tauri::command]
 fn control_save_config(state: State<SharedState>) {
     send_control(
         &state.osc_tx,
@@ -877,281 +608,6 @@ fn control_export_layout(state: State<SharedState>, name: Option<String>) {
         &state.osc_tx,
         OscControlMsg::SendNoArgs {
             address: "/omniphony/control/layout/export".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_audio_sample_rate(state: State<SharedState>, sample_rate: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: "/omniphony/control/audio/sample_rate".to_string(),
-            value: sample_rate.max(0),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_audio_config(state: State<SharedState>, payload: serde_json::Value) {
-    let text = match serde_json::to_string(&payload) {
-        Ok(text) => text,
-        Err(_) => return,
-    };
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/config/audio".to_string(),
-            value: text,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_audio_config_apply(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/config/audio/apply".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_audio_output_device(state: State<SharedState>, output_device: String) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/audio/output_device".to_string(),
-            value: output_device.trim().to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn refresh_output_devices(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/audio/output_devices/refresh".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_config(state: State<SharedState>, payload: serde_json::Value) {
-    let text = match serde_json::to_string(&payload) {
-        Ok(text) => text,
-        Err(_) => return,
-    };
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/config/input".to_string(),
-            value: text,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_config_apply(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/config/input/apply".to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_mode(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(
-        trimmed.as_str(),
-        "bridge" | "pipe_bridge" | "live" | "pipewire" | "pipewire_bridge"
-    ) {
-        return;
-    }
-    let normalized = match trimmed.as_str() {
-        "bridge" => "pipe_bridge",
-        "live" => "pipewire",
-        other => other,
-    };
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/mode".to_string(),
-            value: normalized.to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_backend(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(trimmed.as_str(), "pipewire" | "asio") {
-        return;
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/backend".to_string(),
-            value: trimmed,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_node(state: State<SharedState>, value: String) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/node".to_string(),
-            value: value.trim().to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_description(state: State<SharedState>, value: String) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/description".to_string(),
-            value: value.trim().to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_layout(state: State<SharedState>, value: String) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/layout".to_string(),
-            value: value.trim().to_string(),
-        },
-    );
-}
-
-#[tauri::command]
-fn import_input_layout_from_path(
-    state: State<SharedState>,
-    path: String,
-) -> Result<serde_json::Value, String> {
-    let trimmed = path.trim();
-    if trimmed.is_empty() {
-        return Err("empty layout path".to_string());
-    }
-    layouts::load_layout_file(std::path::Path::new(trimmed))
-        .ok_or_else(|| "failed to parse layout file".to_string())?;
-    let contents =
-        fs::read_to_string(trimmed).map_err(|e| format!("failed to read layout file: {e}"))?;
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/layout".to_string(),
-            value: trimmed.to_string(),
-        },
-    );
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/layout_import".to_string(),
-            value: contents,
-        },
-    );
-    Ok(serde_json::json!({ "path": trimmed }))
-}
-
-#[tauri::command]
-fn control_input_live_channels(state: State<SharedState>, value: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: "/omniphony/control/input/live/channels".to_string(),
-            value: value.max(1),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_sample_rate(state: State<SharedState>, value: i32) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendInt {
-            address: "/omniphony/control/input/live/sample_rate".to_string(),
-            value: value.max(1),
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_format(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(trimmed.as_str(), "f32" | "s16") {
-        return;
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/format".to_string(),
-            value: trimmed,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_clock_mode(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(trimmed.as_str(), "dac" | "pipewire" | "upstream") {
-        return;
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/clock_mode".to_string(),
-            value: trimmed,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_map(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(trimmed.as_str(), "7.1-fixed") {
-        return;
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/map".to_string(),
-            value: trimmed,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_live_lfe_mode(state: State<SharedState>, value: String) {
-    let trimmed = value.trim().to_ascii_lowercase();
-    if !matches!(trimmed.as_str(), "object" | "direct" | "drop") {
-        return;
-    }
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendString {
-            address: "/omniphony/control/input/live/lfe_mode".to_string(),
-            value: trimmed,
-        },
-    );
-}
-
-#[tauri::command]
-fn control_input_apply(state: State<SharedState>) {
-    send_control(
-        &state.osc_tx,
-        OscControlMsg::SendNoArgs {
-            address: "/omniphony/control/input/apply".to_string(),
         },
     );
 }
